Fix macos build for fr-1051
Committer: Hai Huang <haih@google.com>

On branch fr-1051
Your branch is up to date with 'origin/fr-1051'.

Changes to be committed:
modified:   src/benchmark_api_internal.cc
modified:   src/benchmark_api_internal.h
modified:   src/benchmark_runner.cc

+++ src/benchmark_api_internal.cc
@@ -121,7 +121,7 @@ IterationCount BenchmarkInstance::iterations() const {
   return iterations_;
 }
 
-int BenchmarkInstance::repetitions() const {
+size_t BenchmarkInstance::repetitions() const {
   return repetitions_;
 }
